Do not open WebSockets until a client actually sends a message.
The SDK expects to open a WebSocket when the user sends a message,
and close it when the backend has finished sending all responses.

Before this change, when a client created a Streamer, a WebSocket
would be connected, even though the client hadn't necessarily sent
a message.

After this change, creating the Streamer does not open that initial
WebSocket.

This also means that creating a Streamer is no longer async. (This
caused a lot of drama in the unit tests.)

TESTED=updated unit tests, manually tested full e2e system.